chatbot-backend/requirements.txt
Flask==2.0.1
openai
python-dotenv
